Feature changes
-changed account variable from int to long. More coins!
-added trade station village gen
-changed building style for bank and shop
-removed custom card from ATM
-removed enchiridion book since it is outdated.

+++ java/universalcoins/tile/TileBandit.java
@@ -140,7 +140,7 @@ public class TileBandit extends TileEntity implements IInventory {
 		cardAvailable = false;
 		if (inventory[itemCardSlot] != null && inventory[itemCardSlot].hasTagCompound() && !worldObj.isRemote) {
 			String account = inventory[itemCardSlot].getTagCompound().getString("accountNumber");
-			int accountBalance = UniversalAccounts.getInstance().getAccountBalance(account);
+			long accountBalance = UniversalAccounts.getInstance().getAccountBalance(account);
 			if (accountBalance > spinFee) {
 				cardAvailable = true;
 			}
